fastapi-notes/18-dwarpal/requirements.txt
fastapi[standard]
sqlmodel
uvicorn
passlib[bcrypt]
python-jose[cryptography]
python-multipart
pydantic-settings
